requirements.txt
alembic
bcrypt
cryptography
fastapi
psycopg2-binary
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2
langchain
python-dotenv
google-generativeai
langchain-google-genai
Markdown
bs4